Removed the fremData argument from createVPC and calcParameters.
Made it possible to pass a data frame with ext information through createVPC to calcFFEM (instead of alwas reading it in calcVPC and then pass it to calcFFEM).

Made it possible to pass a data frame as dataFile instead of a file name.

Added ... argument to calcParameters to be able to pass arguments to underlying functions.

Made it possible to to pass a data frame as dataFile instead of a file name.